requirements.txt
Pillow==7.0.0
opencv-python==4.2.0.34
openpyxl==3.0.3
scipy==1.4.1
tensorflow==2.1.0
tensorflow-estimator==2.1.0
tensorboard==2.1.1
keras==2.3.1
numpy==1.18.1
mtcnn==0.1.0